Add files via upload
This is a full stack application that I built in IIT Madras, this application lets Sponsors create ad campaigns, search for influencers and request for ads, vice versa influencers can also register themselves,search and request for ads. Further chat assistance is provided between the clients. This application was developed with templates using Jinja2. SQLAlchemy is used as a database toolkit with 4 classes.